De-NSInteger-ize {CG,O2}FontGetGlyphAdvances()
Turns out it didn't just break the build (https://github.com/darlinghq/darling-cocotron/commit/269215f67029441269482261b27289d6f961bd76),
it was actually *wrong*; the Apple's version uses a plain int (not even a CFIndex) for some reason. Do the same in O2FontGetGlyphAdvances(),
but keep the internal ivar as an NSInteger array pointer; we're making a copy in O2FontGetGlyphAdvances() anyway.

+++ AppKit/X11.subproj/O2Context_builtin_FT.m
@@ -189,7 +189,7 @@ static void renderFreeTypeBitmap(
    
    O2PaintRelease(paint);
    
-   NSInteger glyphAdvances[count];
+   int glyphAdvances[count];
    O2Float unitsPerEm=O2FontGetUnitsPerEm(font);
    
    O2FontGetGlyphAdvances(font,glyphs,count,glyphAdvances);
